requirements.txt
pandas
numpy
fpdf2
streamlit
matplotlib
reportlab
seaborn
